commit all recent PR226 changes

+++ multiTDC.h
@@ -1,7 +1,8 @@
 #ifndef __multiTDC__
 #define __multiTDC__ 1
 
-/* This header file defines the 'multiTDC' class. This is a class used to store the information about TDC values for ancillary detectors in K600 experiments, given that the V1190 TDCs can (and do!) run in multihit mode.
+/* This header file defines the 'multiTDC' class. This is a class used to store the information about TDC values 
+ * for ancillary detectors in K600 experiments, given that the V1190 TDCs can (and do!) run in multihit mode.
  * 
  * * Author: Philip Adsley, July 2014
  * E-mail: padsley@gmail.com
@@ -13,7 +14,7 @@
 #include <vector>
 #include <stdio.h>
 
-extern double PulseLimits[2];
+extern int PulseLimits[2];
 
 class multiTDC
 {
